Moved load/unload commands to cog as dpy library doesn't support commands in bot subclass.
